Move DetectCloaked to Mods.Common
