+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -170,7 +170,6 @@
     <Compile Include="Traits\Crates\RevealMapCrateAction.cs" />
     <Compile Include="Traits\Crates\SupportPowerCrateAction.cs" />
     <Compile Include="Traits\DemoTruck.cs" />
-    <Compile Include="DetectCloaked.cs" />
     <Compile Include="Effects\GpsDot.cs" />
     <Compile Include="Effects\Parachute.cs" />
     <Compile Include="Effects\RepairIndicator.cs" />
